requirements.txt
argparse
python-nmap